% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_resid.R
\name{boot_resid}
\alias{boot_resid}
\title{Resampling residual via boostrap}
\usage{
boot_resid(Y, X, x_new, R, seed)
}
\arguments{
\item{Y}{the vector of response}

\item{X}{the matrix of covariates (including intercept)}

\item{x_new}{a new observation of covariates}

\item{R}{the totoal times of resampling}

\item{seed}{the random number generator (RNG) state}
}
\description{
This function return LSE of linear model Y = X beta + epsilon, the fitted value of Y,
the raw residual, the modified residual and the predicted value for new observation x_new,
as well as its 95% confidence interval,
}
\examples{
cat_function()
library(boot)
nuke <- nuclear[, c(1, 2, 5, 7, 8, 10, 11)]
Y <- log(nuke$cost) # True value of Y
X <- data.frame(1,nuke$date,log(nuke$cap),nuke$ne,nuke$ct,log(nuke$cum.n),nuke$pt)
X <- data.matrix(X) # Desgin matrix X
colnames(X) <- NULL 
x_new <- c(1,73,log(886),0,0,log(11),1) # new observation
R <- 1000 # Resampling sample size
seed <- 921 # Random seed - a integer
}
\author{
Jingyi Lin
}
\keyword{Boostrap}
